pyproject.toml
[project]
name = "stitchcv"
version = "0.1.0"
description = "Agentic AI Job Application System"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "playwright>=1.48.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "bcrypt>=4.1.2",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.9",
    "pypdf>=4.0.0",
    "reportlab>=4.2.0",
    "pymupdf>=1.24.0",
    "rapidocr-onnxruntime>=1.3.24",
    "pillow>=10.0.0",
    "python-docx>=1.1.0",
    "langgraph>=0.0.50",
    "langchain-google-genai>=1.0.3",
    "langchain-groq>=0.1.0",
    "langchain-chroma>=0.1.0",
    "langchain-openai>=0.1.0",
    "chromadb>=0.4.0",
    "razorpay>=1.4.1",
    "pgvector>=0.2.5",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
]

[project.scripts]
stitchcv = "stitchcv.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
